remove old typechecker
