+++ src/main/scala/typechecker/TypeCheck.scala
@@ -2,7 +2,6 @@ package fuselang.typechecker
 
 import Subtyping._
 import TypeEnv._
-import Gadgets._
 import fuselang.Utils._
 
 import fuselang.common._
@@ -73,6 +72,11 @@ import Logger.PositionalLoggable
  */
 object TypeChecker {
 
+  def pr[A](v: A): A = {
+    println(v)
+    v
+  }
+
   /* A program consists of a list of function or type definitions, a list of
    * variable declarations and then a command. We build up an environment with
    * all the declarations and definitions, then check the command in that environment
@@ -91,7 +95,7 @@ object TypeChecker {
 
   private def checkDef(defi: Definition, env: Environment) = defi match {
     case FuncDef(id, args, ret, bodyOpt) => {
-      val (env2, _) = env.withScope(1) { newScope =>
+      val env2 = env.withScope { newScope =>
 
         // Bind all declarations to the body.
         val envWithArgs = args.foldLeft(newScope)({ case (env, Decl(id, typ)) =>
@@ -100,17 +104,12 @@ object TypeChecker {
           env.add(id, rTyp)
         })
 
-        // Add physical resources corresponding to array decls
-        val envWithResources = args
-          .collect({ case Decl(id, t:TArray) => id -> t})
-          .foldLeft(envWithArgs)({ case (env, (id, t)) => addPhysicalResource(id, t)(env)})
-
         // Add the return type
-        val envWithRet = envWithResources.withReturn(ret)
+        val envWithRet = envWithArgs.withReturn(ret)
 
         bodyOpt
           .map(body => checkC(body)(envWithRet))
-          .getOrElse(envWithResources)
+          .getOrElse(envWithRet)
       }
       env2.add(id, TFun(args.map(_.typ), ret))
     }
@@ -118,75 +117,6 @@ object TypeChecker {
       val rFields = fields.map({case (k, t) => k -> env.resolveType(t)})
       env.addType(name, TRecType(name, rFields))
     }
-  }
-
-  /**
-   * Add physical resources and default accessor gadget corresponding to a new
-   * array. This is used for `decl` with arrays and new `let` bound arrays.
-   */
-  private def addPhysicalResource(id: Id, typ: TArray)
-                                 (implicit env: Environment) = {
-    val banks = typ.dims.map(_._2)
-    env
-      .addResource(id, banks , typ.ports)
-      .addGadget(id, MultiDimGadget(ResourceGadget(id, banks), typ.dims))
-  }
-
-  /**
-   * Generate a ConsumeList corresponding to the underlying memory type and
-   * the index accessors.
-   */
-  private def getConsumeList(idxs: List[Expr], dims: List[DimSpec])
-                            (implicit arrId: Id, env: Environment) = {
-    val (nEnv, bres, consume) = idxs.zipWithIndex.foldLeft((env, 1, IndexedSeq[Seq[Int]]()))({
-      case ((env1, bres, consume), (idx, dim)) => checkE(idx)(env1) match {
-        // Index is an index type.
-        case (TIndex((s, e), _), env2) =>
-          if ((e - s) % dims(dim)._2 != 0)
-            throw BankUnrollInvalid(arrId, dims(dim)._2, e - s)(idx.pos)
-          else
-            (env2, bres * (e - s), Range(s, e) +: consume)
-        // Index is a statically known number.
-        case (TStaticInt(v), env2) =>
-          (env2, bres * 1, Vector(v % dims(dim)._2) +: consume)
-        // Index is a dynamic number.
-        case (_:TSizedInt, env2) =>
-          if (dims(dim)._2 != 1) throw InvalidDynamicIndex(arrId, dims(dim)._2)
-          else (env2, bres * 1, Vector(0) +: consume)
-
-        case (t, _) =>
-          throw UnexpectedType(idx.pos, "array indexing", "integer type", t)
-      }
-    })
-
-    // Reverse the types list to match the order with idxs.
-    (nEnv, bres, consume.reverse)
-  }
-
-  private def checkLVal(e: Expr)(implicit env: Environment) = e match {
-    case acc@EArrAccess(id, idxs) => env(id).matchOrError(e.pos, "array access", s"array") {
-      // This only triggers for l-values.
-      case TArray(typ, dims, _) => {
-        if (dims.length != idxs.length) {
-          throw IncorrectAccessDims(id, dims.length, idxs.length)
-        }
-        // Bind the type of to Id
-        id.typ = Some(env(id));
-        // Consumption check
-        acc.consumable match {
-          case Some(Annotations.ShouldConsume) => {
-            val (e1, bres, consumeList) = getConsumeList(idxs, dims)(id, env)
-            // Check if the accessors generated enough copies for the context.
-            if (bres != env.getResources)
-              throw InsufficientResourcesInUnrollContext(env.getResources, bres, e)
-            // Consume the resources required by this gadget.
-            typ -> e1.consumeWithGadget(id, consumeList)(acc.pos)
-          }
-          case con => throw Impossible(s"$acc in write position has $con annotation")
-        }
-      }
-    }
-    case _ => checkE(e)
   }
 
   private def checkB(t1: Type, t2: Type, op: BOp) = op match {
@@ -266,8 +196,7 @@ object TypeChecker {
       checkB(t1, t2, op) -> env2
     }
     case EApp(f, args) => env(f) match {
-      case tf@TFun(argTypes, retType) => {
-        f.typ = Some(tf)
+      case TFun(argTypes, retType) => {
         if (argTypes.length != args.length) {
           throw ArgLengthMismatch(expr.pos, argTypes.length, args.length)
         }
@@ -277,16 +206,7 @@ object TypeChecker {
           if (isSubtype(typ, expectedTyp) == false) {
             throw UnexpectedSubtype(arg.pos, "parameter", expectedTyp, typ)
           }
-          // If an array id is used as a parameter, consume it completely.
-          // This works correctly with capabilities.
-          (typ, arg) match {
-            case (ta:TArray, EVar(gadget)) => {
-              val consumeList = ta.dims.map(dim => 0.until(dim._2))
-              e1.consumeWithGadget(gadget, consumeList)(arg.pos)
-            }
-            case (_:TArray, expr) => throw Impossible(s"Type of $expr is $typ")
-            case _ => e1
-          }
+          e1
         }})
       }
       case t => throw UnexpectedType(expr.pos, "application", "function", t)
@@ -298,36 +218,32 @@ object TypeChecker {
       }
       case (t, _) => throw UnexpectedType(expr.pos, "record access", "record type", t)
     }
-    case acc@EArrAccess(id, idxs) => env(id).matchOrError(expr.pos, "array access", s"array type"){
-      // This only triggers for r-values. l-values are checked in checkLVal
+    case EArrAccess(id, idxs) => env(id).matchOrError(expr.pos, "array access", s"array type"){
       case TArray(typ, dims, _) => {
         if (dims.length != idxs.length) {
           throw IncorrectAccessDims(id, dims.length, idxs.length)
         }
+        idxs.foldLeft(env)((env, idx) => {
+          val (typ, nEnv) = checkE(idx)(env)
+          typ match {
+            case _:IntType => ()
+            case _ => throw UnexpectedType(idx.pos, "array index", "integer type", typ)
+          }
+          nEnv
+        })
         // Bind the type of to Id
         id.typ = Some(env(id));
-        // Consumption check
-        acc.consumable match {
-          case None => throw Impossible(s"$acc in read position has no consumable annotation")
-          case Some(Annotations.SkipConsume) => typ -> env
-          case Some(Annotations.ShouldConsume) => {
-            val (e1, _, consumeList) = getConsumeList(idxs, dims)(id, env)
-            // Consume the resources required by this gadget.
-            typ -> e1.consumeWithGadget(id, consumeList)(acc.pos)
-          }
-        }
+        typ -> env
       }
     }
   }
 
-  /**
-   * Checks a given simple view and returns the dimensions for the view,
-   * shrink factors, and an updated environment.
-   */
+  // Check if this array dimension is well formed and return the dimension
+  // spec for the corresponding dimension in the view.
   private def checkView(view: View, arrDim: DimSpec)
-                       (implicit env: Environment): (Environment, Int, DimSpec) = {
+                       (implicit env: Environment): (Environment, DimSpec) = {
 
-    val View(suf, pre, shrink) = view
+    val View(suf, prefix, shrink) = view
     val (len, bank) = arrDim
 
     // Shrinking factor must be a factor of banking for the dimension
@@ -337,22 +253,26 @@ object TypeChecker {
 
     val newBank = shrink.getOrElse(bank)
 
+    // Get the indexing expression
     val idx = suf match {
-      case Aligned(fac, idx) => if (newBank > fac || fac % newBank != 0) {
-        throw InvalidAlignFactor(suf.pos, "Invalid Align")
+      case Aligned(fac, idx) => if (newBank > fac) {
+        throw InvalidAlignFactor(suf.pos, s"Invalid align factor. Banking factor $newBank is bigger than alignment factor $fac.")
+      }
+      else if(fac % newBank != 0) {
+        throw InvalidAlignFactor(suf.pos,
+          s"Invalid align factor. Banking factor $newBank not a factor of the alignment factor $fac.")
       } else {
         idx
       }
       case Rotation(idx) => idx
     }
 
-    // Check the index is non-index int type
-    val (typ, nEnv) = checkE(idx)(env)
+    val (typ, nEnv) = checkE(idx)
     typ.matchOrError(idx.pos, "view", "integer type") {
       case _:IntType => () // IntTypes are valid
     }
 
-    (nEnv, newBank, (pre.getOrElse(len) -> newBank))
+    (nEnv, (prefix.getOrElse(len) -> newBank))
   }
 
   private def checkPipeline(enabled: Boolean, loop: Command, body: Command) = {
@@ -368,12 +288,14 @@ object TypeChecker {
   private def checkC(cmd: Command)
                     (implicit env:Environment): Environment = cmd match {
     case CPar(c1, c2) => checkC(c2)(checkC(c1))
+    case CSeq(c1, c2) => checkC(c2)(checkC(c1))
     case CIf(cond, cons, alt) => {
       val (cTyp, e1) = checkE(cond)(env)
       cTyp.matchOrError(cond.pos, "if condition", "bool"){ case _:TBool => () }
-      val (e2, _) = e1.withScope(1)(e => checkC(cons)(e))
-      val (e3, _) = e1.withScope(1)(e => checkC(alt)(e))
-      e2 merge e3
+      e1.withScope(e => checkC(cons)(e))
+      e1.withScope(e => checkC(alt)(e))
+      // No binding updates need to be reflected.
+      e1
     }
     case CWhile(cond, pipeline, body) => {
       checkPipeline(pipeline, cmd, body)
@@ -381,29 +303,20 @@ object TypeChecker {
       if (cTyp != TBool()) {
         throw UnexpectedType(cond.pos, "while condition", TBool().toString, cTyp)
       }
-      e1.withScope(1)(e => checkC(body)(e))._1
+      e1.withScope(e => checkC(body)(e))
     }
     case CUpdate(lhs, rhs) => {
-      val (t1, e1) = checkLVal(lhs)
+      val (t1, e1) = checkE(lhs)
       val (t2, e2) = checkE(rhs)(e1)
       if (isSubtype(t2, t1)) e2
       else throw UnexpectedSubtype(rhs.pos, "assignment", t1, t2)
     }
-    case CReduce(rop, l, r) => {
-      val (t1, e1) = checkLVal(l)
-      val (ta, e2) = checkE(r)(e1)
-      (t1, ta) match {
-        case (t1, TArray(t2, dims, _)) =>
-          if (isSubtype(t2, t1) == false ||
-              dims.length != 1 ||
-              dims(0)._1 != dims(0)._2) {
-            throw UnexpectedType(r.pos, s"reduction operator $rop", "fully banked array", ta)
-          } else {
-            e2
-          }
-        case _ =>
-          throw UnexpectedType(r.pos, s"reduction operator $rop", "fully banked array", ta)
-      }
+    case CReduce(_, l, r) => {
+      val (t1, e1) = checkE(l)
+      val (t2, e2) = checkE(r)(e1)
+
+      if (isSubtype(t2, t1)) e2
+      else throw UnexpectedSubtype(r.pos, "reduction operator", t1, t2)
     }
     case l@CLet(id, typ, Some(EArrLiteral(idxs))) => {
       val expTyp = typ.getOrThrow(ExplicitTypeMissing(l.pos, "Array literal", id))
@@ -426,9 +339,7 @@ object TypeChecker {
             })
 
             id.typ = typ
-
-            // Add the type binding, physical resource, and the accessor.
-            addPhysicalResource(id, ta)(nEnv).add(id, expTyp)
+            nEnv.add(id, ta)
           }
         }
     }
@@ -479,7 +390,9 @@ object TypeChecker {
         case _:TRational => TDouble()
         case t => t
       })
+      // Check the type of the expression
       val (t, e1) = checkE(exp)
+      // Check if type of expression is a subtype of the annotated type.
       rTyp match {
         case Some(t2) => {
           if (isSubtype(t, t2))
@@ -505,71 +418,49 @@ object TypeChecker {
                                         "Let binding without initializer",
                                         id))
 
-      // If this is an array literal, bind the physical resource too.
-      val nEnv = fullTyp match {
-        case ta: TArray => addPhysicalResource(id, ta)(env)
-        case _ => env
-      }
-
-      nEnv.add(id, fullTyp)
+      env.add(id, fullTyp)
     }
     case CFor(range, pipeline, par, combine) => {
+      // Check if this is a pipelined loop.
       checkPipeline(pipeline, cmd, par)
 
-      val iter = range.iter
-      val (e1, binds) = env.withScope(range.u) { newScope =>
+      // Check the for loop body and make variables declared in loop body
+      // available to the combine block.
+      env.withScope { newScope =>
         // Add binding for iterator in a separate scope.
-        val e2 = newScope.add(iter, range.idxType)
-        checkC(par)(e2)
+        val forEnv = newScope.add(range.iter, range.idxType)
+        val combEnv = checkC(par)(forEnv)
+        checkC(combine)(combEnv)
       }
-      // Create scope where ids bound in the parallel scope map to fully banked
-      // arrays. Remove the iterator binding.
-      val vecBinds = binds
-        .withFilter({ case (id, _) => id != iter })
-        .map({ case (id, typ) =>
-          id -> TArray(typ, List((range.u, range.u)), 1)
-        })
-
-      e1.withScope(1)(e2 => checkC(combine)(e2 ++ vecBinds))._1
     }
-    case view@CView(id, arrId, vdims) => env(arrId) match {
+    case CView(id, arrId, vdims) => env(arrId) match {
       case TArray(typ, adims, port) => {
-        // Check if the view is defined inside an unroll
-        if (env.getResources != 1) {
-          throw ViewInsideUnroll(view.pos)
-        }
-
         val (vlen, alen) = (vdims.length, adims.length)
         if(vlen != alen) {
           throw IncorrectAccessDims(arrId, alen, vlen)
         }
 
-        // Check all dimensions in the view are well formed.
-        val (env1, shrinks, ndims) = adims.zip(vdims).foldLeft((env, List[Int](), List[DimSpec]()))({
-          case ((env, shrinks, dims), (arrDim, view)) =>
-            val (nEnv, shrink, dim) = checkView(view, arrDim)(env)
-            (nEnv, shrink :: shrinks, dim :: dims)
-        })
+        val (env1, viewDims) =
+          adims
+            .zip(vdims)
+            .foldLeft((env, List[DimSpec]()))({
+              case ((env, viewDims), (arrDim, view)) =>
+                val (nEnv, vDim) = checkView(view, arrDim)(env)
+                (nEnv, vDim :: viewDims)
+            })
 
-        // Fully consume the array
-        val nEnv = env1.addGadget(id, viewGadget(env1.getGadget(arrId), shrinks.reverse, adims))
-
-        // Annotate the ids in the expressions
-        val viewTyp = TArray(typ, ndims.reverse, port)
+        // View get the same type the array they are built from.
+        val viewTyp = TArray(typ, viewDims.reverse, port)
         id.typ = Some(viewTyp)
         arrId.typ = Some(env(arrId))
 
-        // Add binding for the new array. The dimensions are reversed because
-        // we used foldLeft above.
-        nEnv.add(id, viewTyp)
+        // Add binding for the new array
+        env1.add(id, viewTyp)
       }
       case t => throw UnexpectedType(cmd.pos, "view", "array", t)
     }
-    case view@CSplit(id, arrId, dims) => env(arrId) match {
+    case CSplit(id, arrId, dims) => env(arrId) match {
       case TArray(typ, adims, ports) => {
-        if (env.getResources != 1) {
-          throw ViewInsideUnroll(view.pos)
-        }
         val (vlen, alen) = (dims.length, adims.length)
         if(vlen != alen) {
           throw IncorrectAccessDims(arrId, alen, vlen)
@@ -594,38 +485,23 @@ object TypeChecker {
           case ((dim, bank), n) => throw InvalidSplitFactor(id, arrId, n, bank, dim)
         })
 
-        // Create a gadget for the view
-        val nEnv = env.addGadget(id,
-          splitGadget(env.getGadget(arrId), adims , viewDims))
-
         // Annotate the ids in the expressions
         val viewTyp = TArray(typ, viewDims, ports)
         id.typ = Some(viewTyp)
         arrId.typ = Some(env(arrId))
 
-        nEnv.add(id, viewTyp)
+        env.add(id, viewTyp)
       }
       case t => throw UnexpectedType(cmd.pos, "split", "array", t)
     }
-    case CSeq(c1, c2) => {
-      val (env1, binds) = env.withScope(1) { newScope =>
-        checkC(c1)(newScope)
-      }
-      val env2 = checkC(c2)(env ++ binds)
-      env2 merge env1
-    }
     case CExpr(e) => checkE(e)._2
     case CReturn(expr) => {
-      env.getReturn match {
-        case Some(retType) => {
-          val (t, e) = checkE(expr)
-          if (isSubtype(t, retType) == false) {
-            throw UnexpectedSubtype(expr.pos, "return", retType, t)
-          }
-          e
-        }
-        case None => throw ReturnNotInFunc(cmd.pos)
+      val retType = env.getReturn.get
+      val (t, e) = checkE(expr)
+      if (isSubtype(t, retType) == false) {
+        throw UnexpectedSubtype(expr.pos, "return", retType, t)
       }
+      e
     }
     case CEmpty => env
     case _:CDecorate => env
